chore(checkov): refactor files structure
