+++ src/checkov/checkovRunner.ts
@@ -3,44 +3,11 @@ import { spawn } from 'child_process';
 import * as path from 'path';
 import { Logger } from 'winston';
 import { CheckovInstallation } from './checkovInstaller';
-import { convertToUnixPath, getGitRepoName, getDockerPathParams, runVersionCommand, normalizePath } from './utils';
+import { convertToUnixPath, getGitRepoName, getDockerPathParams, runVersionCommand, normalizePath } from '../utils';
+import { CheckovResponse, CheckovResponseRaw } from './models';
+import { parseCheckovResponse } from './checkovParser';
 
-export interface FailedCheckovCheck {
-    checkId: string;
-    checkName: string;
-    fileLineRange: [number, number];
-    resource: string;
-    guideline?: string;
-    fixedDefinition?: string;
-}
 
-interface CheckovResponse {
-    results: {
-        failedChecks: FailedCheckovCheck[];
-    };
-}
-
-interface SuccessResponseRaw {
-    resource_count: 0;
-    results: undefined;
-}
-
-interface FailedCheckovCheckRaw {
-    check_id: string;
-    bc_check_id: string | undefined;
-    check_name: string;
-    file_line_range: [number, number];
-    resource: string;
-    guideline?: string;
-    description?: string;
-    fixed_definition?: string;
-}
-
-interface CheckovResponseRaw {
-    results: {
-        failed_checks: FailedCheckovCheckRaw[];
-    };
-}
 
 const dockerMountDir = '/checkovScan';
 const configMountDir = '/checkovConfig';
@@ -158,37 +125,3 @@ export const runCheckovScan = (logger: Logger, checkovInstallation: CheckovInsta
     });
 };
 
-const parseCheckovResponse = (rawResponse: CheckovResponseRaw | SuccessResponseRaw, useBcIds: boolean | undefined): CheckovResponse => {
-    if (!(Array.isArray(rawResponse) || rawResponse.results)) {
-        if  (rawResponse.resource_count === 0) {
-            return {
-                results: {
-                    failedChecks: []
-                }
-            };
-        } else {
-            throw new Error('Unexpected checkov response');
-        }
-    }
-
-    let failedChecks: FailedCheckovCheckRaw[];
-    if (Array.isArray(rawResponse)) {
-        failedChecks = rawResponse.reduce((res, val) => res.concat(val.results.failed_checks), []);
-    }
-    else {
-        failedChecks = rawResponse.results.failed_checks;
-    }
-
-    return {
-        results: {
-            failedChecks: failedChecks.map(rawCheck => ({
-                checkId: (useBcIds && rawCheck.bc_check_id) || rawCheck.check_id,
-                checkName: rawCheck.check_name,
-                fileLineRange: rawCheck.file_line_range,
-                resource: rawCheck.resource,
-                guideline: rawCheck.guideline || rawCheck.description,
-                fixedDefinition: rawCheck.fixed_definition
-            }))
-        }
-    };
-};
