+++ src/checkov/index.ts
@@ -1,0 +1,3 @@
+export { FailedCheckovCheck } from './models';
+export { runCheckovScan } from './checkovRunner';
+export { CheckovInstallation, installOrUpdateCheckov } from './checkovInstaller';
