+++ src/diagnostics.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import { FailedCheckovCheck } from './checkovRunner';
+import { FailedCheckovCheck } from './checkov';
 
 export interface DiagnosticReferenceCode {
     target: vscode.Uri;
@@ -24,7 +24,8 @@ export const applyDiagnostics = (document: vscode.TextDocument, diagnostics: vsc
             range: new vscode.Range(startPos, line.range.end),
             severity: vscode.DiagnosticSeverity.Error,
             source: 'Checkov ',
-            code
+            code,
+
         });
     }
